import scripts code optimization
